Import evm-providers from package instead evm package
