+++ packages/xchain-ethereum/package.json
@@ -39,6 +39,7 @@
     "@xchainjs/xchain-crypto": "^0.3.0",
     "@xchainjs/xchain-evm": "^0.2.2",
     "@xchainjs/xchain-util": "^0.13.0",
+    "@xchainjs/xchain-evm-providers": "^0.1.0",
     "axios": "^1.3.6",
     "ethers": "^5.7.2"
   },
@@ -47,6 +48,7 @@
     "@xchainjs/xchain-crypto": "^0.3.0",
     "@xchainjs/xchain-evm": "^0.2.2",
     "@xchainjs/xchain-util": "^0.13.0",
+    "@xchainjs/xchain-evm-providers": "^0.1.0",
     "axios": "^1.3.6",
     "ethers": "^5.7.2"
   }
