+++ packages/xchain-ethereum/src/const.ts
@@ -1,5 +1,6 @@
 import { ExplorerProvider, Network } from '@xchainjs/xchain-client'
-import { EVMClientParams, EtherscanProvider } from '@xchainjs/xchain-evm'
+import { EVMClientParams } from '@xchainjs/xchain-evm'
+import { EtherscanProvider } from '@xchainjs/xchain-evm-providers'
 import { Asset } from '@xchainjs/xchain-util'
 import { BigNumber, ethers } from 'ethers'
