+++ packages/xchain-evm/package.json
@@ -37,6 +37,7 @@
     "@xchainjs/xchain-client": "^0.14.1",
     "@xchainjs/xchain-crypto": "^0.3.0",
     "@xchainjs/xchain-util": "^0.13.0",
+    "@xchainjs/xchain-evm-providers": "^0.1.0",
     "axios": "^1.3.6",
     "ethers": "^5.7.2"
   },
